The very raw begginnigs of something
Most of the work here is to upgrade the build system to something that's
not a hand crafted Makefile. I picked CMake because it seems popular and
I didn't want to go back to Autotools.

As for the changes to source this doesn't work at all. It compiles and
probably even still runs. However I don't have anything written yet to
start testing the various aspects. For that I'm contemplating generating
a python client that we can then use via nose2/pytest to exercise the
code base.